+++ src/tests/hook.test.ts
@@ -4,10 +4,12 @@ import {
 import * as hookService from '../services/hookService';
 import * as panelService from '../services/panelService';
 import * as panelSetService from '../services/panelSetService';
+import * as panelSet from '../requestHandlers/panelSet';
 import { Sequelize } from 'sequelize';
 jest.mock('../services/hookService');
 jest.mock('../services/panelService');
 jest.mock('../services/panelSetService');
+jest.mock('../requestHandlers/panelSet');
 
 const sequelizeMock = () => ({} as jest.Mocked<Sequelize>);
 
@@ -99,9 +101,10 @@ describe('Create Hook Controller', () => {
         };
 
         (panelService.getPanel as jest.Mock).mockReturnValue(() => Promise.resolve(panelData));
+        (panelSet.validateHookConnection as jest.Mock).mockReturnValue(() => Promise.resolve());
         (hookService.createHook as jest.Mock).mockReturnValue(() => Promise.resolve(hookData));
 
-        const response = await _createHookController(sequelizeMock())(JSON.parse(`[1, 1]`), 1, 2);
+        const response = await _createHookController(sequelizeMock())(JSON.parse(`[1, 1]`), 1, 2, false);
 
         expect(response).toBe(hookData);
     });
@@ -111,15 +114,24 @@ describe('Create Hook Controller', () => {
 
         (panelService.getPanel as jest.Mock).mockReturnValue(() => Promise.resolve(null));
 
-        const response = await _createHookController(sequelizeMock())(JSON.parse(`[1, 1]`), 1, 2);
+        const response = await _createHookController(sequelizeMock())(JSON.parse(`[1, 1]`), 1, 2, false);
 
         expect(response).toEqual(error);
     });
 
+    test('If an error occurs in validate, error should be returned', async () => {
+        (panelSet.validateHookConnection as jest.Mock).mockReturnValue(() => {throw new Error('Connection already exists.')});
+
+        const response = await _createHookController(sequelizeMock())(JSON.parse(`[1, 1]`), 1, 2, false);
+
+        expect(response).toBeInstanceOf(Error);
+    });
+
     test('If an error occurs, error should be returned', async () => {
+        (panelSet.validateHookConnection as jest.Mock).mockReturnValue(() => Promise.resolve());
         (hookService.createHook as jest.Mock).mockReturnValue(() => { throw new Error('Error Message'); });
 
-        const response = await _createHookController(sequelizeMock())(JSON.parse(`[1, 1]`), 1, 2);
+        const response = await _createHookController(sequelizeMock())(JSON.parse(`[1, 1]`), 1, 2, false);
 
         expect(response).toBeInstanceOf(Error);
     });
@@ -134,31 +146,32 @@ describe('Add Set To Hook Controller', () => {
         };
         const panelSetData = { author_id: 'abc123-efg456-hij789' };
 
-        (panelSetService.getPanelSetByID as jest.Mock).mockReturnValue(() => Promise.resolve(panelSetData));
+        (panelSet.validateHookConnection as jest.Mock).mockReturnValue(() => Promise.resolve());
         (hookService.addSetToHook as jest.Mock).mockReturnValue(() => Promise.resolve(hookData));
 
-        const response = await _addSetToHookController(sequelizeMock())(1, 2);
+        const response = await _addSetToHookController(sequelizeMock())(1, 2, false);
 
         expect(response).toBe(hookData);
     });
 
-    test('If panel set does not exist, return panel set error', async () => {
-        const error = new Error('no panel_set exists for given panel_set_id');
+    test('If an error occurs in validate, the error should be returned', async () => {
+        const panelSetData = { author_id: 'abc123-efg456-hij789' };
 
-        (panelSetService.getPanelSetByID as jest.Mock).mockReturnValue(() => Promise.resolve(null));
+        (panelSet.validateHookConnection as jest.Mock).mockReturnValue(() => { throw new Error('Error Messgage'); });
 
-        const response = await _addSetToHookController(sequelizeMock())(1, 2);
+        const response = await _addSetToHookController(sequelizeMock())(1, 2, false);
 
-        expect(response).toEqual(error);
+        expect(response).toBeInstanceOf(Error);
     });
+
 
     test('If an error occurs, the error should be returned', async () => {
         const panelSetData = { author_id: 'abc123-efg456-hij789' };
 
-        (panelSetService.getPanelSetByID as jest.Mock).mockReturnValue(() => Promise.resolve(panelSetData));
+        (panelSet.validateHookConnection as jest.Mock).mockReturnValue(() => Promise.resolve());
         (hookService.addSetToHook as jest.Mock).mockReturnValue(() => { throw new Error('Error Messgage'); });
 
-        const response = await _addSetToHookController(sequelizeMock())(1, 2);
+        const response = await _addSetToHookController(sequelizeMock())(1, 2, false);
 
         expect(response).toBeInstanceOf(Error);
     });
